requirements.txt
# one-piece-scan-bot
pyquery==1.2.13
python-telegram-bot==8.1.1
urlfetch==1.1.3
dropbox==11.20.0
selenium==3.141.0
